refactor category directories
