+++ app/src/main/java/com/baghdad/tudee/ui/screens/categoryScreen/CategoryUiState.kt
@@ -1,4 +1,4 @@
-package com.baghdad.tudee.ui.viewModel.state
+package com.baghdad.tudee.ui.screens.categoryScreen
 
 import androidx.annotation.DrawableRes
