+++ app/src/main/java/com/baghdad/tudee/ui/screens/categoryScreen/mapper/PredefindCategoryImageMapper.kt
@@ -1,4 +1,4 @@
-package com.baghdad.tudee.ui.viewModel.mapper
+package com.baghdad.tudee.ui.screens.categoryScreen.mapper
 
 import com.baghdad.tudee.R
 import com.baghdad.tudee.domain.entity.Category
